chore(cli): replace lipgloss with coder/pretty
This change will improve over CLI performance and "snapiness" as well as
substantially reduce our test times (benchmark TODO).

The inefficiency of lipgloss disproportionately impacts our system, as all help
text for every command is generated whenever any command is invoked.

The `pretty` API could clean up a lot of the code (e.g., by replacing
complex string concatenations with Printf), but this commit is too
expansive as is so that work will be done in a follow up.