feat: increase autostop deadline if it passes autostart
If the autostop deadline passes the next autostart deadline, increase
the autostop deadline to be `next_autostart + ttl` instead of `now +
ttl`.

This applies both on workspace build and periodically via the lifecycle
executor system. It cannot happen immediately on activity bump as
there's no way to execute cron statements in the database.